RegistrationForm
I just upgraded it to a next level. It looks awesome now and also it can save the data
given by the user in a separate text file which will be very easy to read and access.
I spend hours and hours completing this project, I hope you will like this one.
That's all thanks.

It's basically a Registration form for a Dance Studio.